Get URLs to daughter pages from a single web page, or multiple pages, on FNA.

Get URLs to daughter pages from a single web page, or multiple pages, on FNA.

parse_page(url, from = "fna", cores, no_cores)

Arguments

url
The URL of the page

from
one of fna or jepson

cores
Use parallel processing in plyr functions (default to FALSE).

no_cores
Number of cores to use in the parallel plyr work.

Value

Taxa names and taxon IDs in a data.frame.

Details

If you use parallelization with argument cores=TRUE, make sure to install doMC first.

Examples

if (FALSE) {
pg1<-'http://www.efloras.org/browse.aspx?flora_id=1&start_taxon_id=10074&page=1'
pg2<-'http://www.efloras.org/browse.aspx?flora_id=1&start_taxon_id=10074&page=2'
pg3<-'http://www.efloras.org/browse.aspx?flora_id=1&start_taxon_id=10074&page=3'
parse_page(pg1)
parse_page(list(pg1, pg2, pg3))
}